backend/requirements.txt
fastapi==0.115.0
uvicorn==0.38.0
pydantic==2.9.2
numpy==2.3.0
pandas==2.3.3
openpyxl==3.1.5
pyinstaller==6.10.0
pytest==8.3.3
httpx==0.27.2
pywin32==306
msoffcrypto-tool==5.4.2
python-multipart>=0.0.9,<0.1.0
jinja2==3.1.6
pydantic-settings
sqlmodel
